Revert "ARM: dts: imx8mm: Replace deprecated fsl,usbphy DT props with phys"
This reverts commit aded940f4fb7e2af6a388e20f1bfb09410962566.

This is a patch that is introduced also in upstrem linux kernel coming
with v5.15. But this also leads to problems with using this binding with
older kernel versions. E.g. if you want to boot a distro like debian,
fedora .. that has an old kernel.

A fixes tag is added to be able to merge it back:
https://lore.kernel.org/all/20210920213224.646520-1-festevam@gmail.com/