kontron-sl-mx8mm: add fixes for backward compatibility
Re-add the changed 'fsl.usbphy' node to prevent older kernels from crash [1]

Debian 11 has no module enabled to support the PC9450 regulator. With
this it is not possible to probe the mmc0 because there is a reference
to the vmmc-supply/vqmmc-supply. Since it is not required for operation
remove this nodes to make mmc0 (eMMC) available.

[1] https://lore.kernel.org/all/20210921113754.767631-1-festevam@gmail.com/

Signed-off-by: Heiko Thiery <heiko.thiery@gmail.com>